feat: Modified Actions, Forms and Authentication

+++ src/actions/deleteTodo.ts
@@ -1,20 +1,24 @@
 "use server";
 
+import { auth } from "@/lib/auth";
 import prisma from "@/lib/db";
 import { revalidatePath } from "next/cache";
 
 export const deleteTodo = async (id: string) => {
+  const session = await auth();
+
+  if (!session) return { success: false, error: "Unauthorized" };
+
   try {
     await prisma.todo.delete({
       where: {
         id: id,
+        userId: session.user.id,
       },
     });
-
-    revalidatePath("/");
-
-    return { success: true };
   } catch (error) {
-    return { success: true, error };
+    return { success: false, error: "Failed to delete todo" };
   }
+
+  revalidatePath("/");
 };
